Add List and change Read endpoint. Change homepage to use endpoints.
A script was run to add each student in students.json to the server/students folder.
A List endpoint was created for api v1 that lists the files in that directory.
The read enpoint was changed to take the name of the file to fetch.

Frontend was changed to hit List endpoint then hit first 10 Read endpoints to load first 10 students.
render() was renamed finishRender() and was shortened; it gets called after 10 students are fetched.
populateTable() became initTable(), and will create the thead and an empty tbody.

In index.css, 'visibility: hidden' was removed from #student-tiles because it was preventing the
tile view from being rendered consistently. Visibility is now handled only by showView().